Fix filesize formating on tests and avoid duplicated function
