+++ .env
@@ -1,4 +1,4 @@
-DJANGO_SECRET_KEY=local
+DJANGO_SECRET_KEY=local-dev-secret-key-not-for-production
 DJANGO_DEBUG=True
 USE_GRANIAN=False
 USE_POSTGRES=False
