+++ src/config/jinja2.py
@@ -4,21 +4,7 @@ from django.urls import reverse
 from django.utils import translation
 from jinja2 import Environment
 
-
-def filesizeformat(value):
-    """Format a file size in bytes to a human-readable string."""
-    try:
-        size = float(value)
-    except (TypeError, ValueError):
-        return "0 B"
-    if size < 1024:
-        return f"{int(size)} B"
-    elif size < 1024 * 1024:
-        return f"{size / 1024:.1f} KB"
-    elif size < 1024 * 1024 * 1024:
-        return f"{size / (1024 * 1024):.1f} MB"
-    else:
-        return f"{size / (1024 * 1024 * 1024):.1f} GB"
+from core.utils import filesizeformat
 
 
 def environment(**options):
