+++ src/core/urls.py
@@ -9,7 +9,6 @@ from core.views.static_views import (
     home_new,
     home_old,
     manifest,
-    marker_generator,
     me_hotsite,
     robots_txt,
     service_worker,
@@ -56,7 +55,6 @@ urlpatterns = [
     path("exhibits/edit-ar/", create_or_edit_ar_exhibit, name="edit-ar-exhibit"),
     path("exhibits/create-mr/", create_or_edit_mr_exhibit, name="create-mr-exhibit"),
     path("exhibits/edit-mr/", create_or_edit_mr_exhibit, name="edit-mr-exhibit"),
-    path("generator/", marker_generator, name="marker-generator"),
     path("marker/", marker_preview, name="marker-preview"),
     path("markers/edit/", edit_marker, name="edit-marker"),
     path("markers/upload/", marker_upload, name="marker-upload"),
